Excluding the ramdisk from GKI boot images
By default, the build system assumes the ramdisk is either
in the boot.img or in the init_boot.img. In gki_arm64, we've
removed init_boot.img so the ramdisk will be included in the
boot.img.

However, GKI boot images should not include a ramdisk.
Sets BOARD_BUILD_SYSTEM_ROOT_IMAGE to true to exclude
the ramdisk.

Bug: 220834917
Test: lunch gki_arm64-user; make -j dist; then unpack_bootimg boot-*.img

Change-Id: I441592c141d6cfdb19a13c98cfca9a976af12604